視点移動処理
プレイヤー移動とともに、視点移動に伴って仮の銃が動けるようにしました。
