+++ game/Assets/Script/Soma.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d17897fa201e8dc42b4b5df9e5d8bd52
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
