+++ game/Assets/Script/Soma/GunController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 50482f47fbde9d94fbc7d93466b1678f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
